Revert "feat: packaging litellm-client for single-shot llm calls"
This reverts commit 6565a5767c258a36321d6bdddc4f174753a59f9b.
