+++ apps/xyne-claw/package.json
@@ -20,7 +20,6 @@
     "@modelcontextprotocol/sdk": "^1.28.0",
     "@sinclair/typebox": "^0.34.0",
     "@upstash/context7-mcp": "^2.1.7",
-    "@xyne/litellm-client": "workspace:*",
     "deepwiki-mcp": "^0.0.6",
     "exceljs": "^4.4.0",
     "express": "^4.21.0",
