+++ apps/xyne-claw/src/twin-soul-synthesizer.ts
@@ -11,7 +11,7 @@
  * never invents — and respects a hard char cap so the file can't blow context.
  */
 
-import { fetchLiteLLMWithRetry } from "@xyne/litellm-client";
+import { fetchLiteLLMWithRetry } from "./litellm-retry.js";
 import { createLogger } from "./logger.js";
 
 const log = createLogger("twin-soul-synthesizer");
