+++ docs/setup/local-development.md
@@ -38,7 +38,7 @@ cd xyne-spaces
 pnpm install
 ```
 
-One install covers all 21 workspace packages. There is no per-package install step —
+One install covers all 20 workspace packages. There is no per-package install step —
 if you find yourself running `pnpm install` inside `apps/backend`, something is wrong.
 
 ## 2. Environment files
@@ -91,14 +91,12 @@ secrets you have added are never overwritten.
 pnpm run build:shared
 ```
 
-This compiles `@xyne/shared`, `@xyne/icons`, `agentic-framework`, and
-`@xyne/litellm-client`. **Do this before anything else** — the applications import
-these from their built `dist/`, so a missing build shows up as unresolved-import
-errors rather than a clear message.
+This compiles `@xyne/shared`, `@xyne/icons`, and `agentic-framework`. **Do this before
+anything else** — the backend and dashboard import these from their built `dist/`, so
+a missing build shows up as unresolved-import errors rather than a clear message.
 
 Re-run it whenever you change code in `packages/shared`, `packages/icons`, or
-`packages/framework`, or `packages/litellm-client`. For an ongoing loop, watch
-instead:
+`packages/framework`. For an ongoing loop, watch instead:
 
 ```bash
 pnpm --filter @xyne/shared run watch
