+++ package.json
@@ -7,7 +7,7 @@
     "prepare": "husky",
     "setup": "pnpm install && pnpm run build:shared",
     "install:all": "pnpm install",
-    "build:shared": "pnpm --filter @xyne/shared --filter @xyne/icons --filter agentic-framework --filter @xyne/litellm-client run build",
+    "build:shared": "pnpm --filter @xyne/shared --filter @xyne/icons --filter agentic-framework run build",
     "services": "chmod +x scripts/start-services.sh && ./scripts/start-services.sh",
     "services:win": "bash scripts/start-services-win.sh",
     "secrets": "node scripts/generate-local-secrets.mjs",
