--- packages/litellm-client/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@xyne/litellm-client",
-  "version": "0.1.0",
-  "description": "Shared LiteLLM HTTP transport for Xyne Node.js services",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "prepare": "npm run build",
-    "test": "npm run build && node --test test/*.test.mjs",
-    "typecheck": "tsc --noEmit"
-  },
-  "engines": {
-    "node": ">=18.17.0"
-  },
-  "dependencies": {
-    "undici": "6.27.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.19.0",
-    "typescript": "^5.7.0"
-  }
-}
